#71 feat: intersection observer 로직 훅으로 분리, useFileCache 로직 분리 심화
