+++ app/main/index.tsx
@@ -14,6 +14,7 @@ import useFileCaches, {
     FeedsCache,
     getLastPageOfConsecutiveList,
 } from "./hooks/useFileCaches";
+import useObserveElement from "./hooks/useObserveElement";
 
 export interface ParsedFeedsDataType {
     id: string;
@@ -90,17 +91,34 @@ export default function MainPage({
     isLocal,
 }: Readonly<MainProps>) {
     const { getDataFrom } = new RequestControllers();
+
     const [currentSort, setCurrentSort] = useState(0);
     const [isFilterFavorite, setIsFilterFavorite] = useState<boolean>(false);
-    const [observerElement, setObserverElement] =
-        useState<HTMLDivElement | null>(null);
     const [totalCount, setTotalCount] = useState<number>(0);
     const [isMobileLayout, setIsMobileLayout] = useState<boolean>(false);
     const [currentPage, setCurrentPage] = useState<number>(1);
     const [renewState, setRenewState] = useState<string>(
         STATE_MESSAGE_STRINGS.start
     );
+    const [feedsToDisplay, setFeedsToDisplay] = useState<ParsedFeedsDataType[]>(
+        []
+    );
+
+    const queryParameters = useRef<string>("");
+    const enabledFilters = useRef<
+        ("favorite" | "source" | "texts" | "sorts")[]
+    >([]);
     const abortControllerRef = useRef<AbortController | null>(null);
+
+    const {
+        cacheContainer,
+        initializeCache,
+        initializeFilteredCache,
+        updateFeedsCache,
+    } = useFileCaches({
+        totalFeedsCount: totalCount,
+        currentPage,
+    });
     const [sourceDisplayState, setSourceDisplayState] = useFilters(
         sources,
         true
@@ -109,6 +127,21 @@ export default function MainPage({
         JSON.stringify(Object.values(SEARCH_OPTIONS)),
         ""
     );
+
+    const updateCurrentPage = (value: number | ((value: number) => number)) => {
+        setCurrentPage(value);
+    };
+
+    const detectIfMobileLayout = useCallback(() => {
+        if (document.documentElement.offsetWidth >= 768) {
+            setIsMobileLayout(false);
+        } else {
+            setIsMobileLayout(true);
+        }
+    }, []);
+
+    useResizeEvent(detectIfMobileLayout, true, [detectIfMobileLayout]);
+
     const { data: searchEnginesList } = useQuery({
         queryKey: [`/search_engines?userId=${userId}`],
         queryFn: () =>
@@ -119,38 +152,6 @@ export default function MainPage({
                   ),
     });
 
-    const [feedsToDisplay, setFeedsToDisplay] = useState<ParsedFeedsDataType[]>(
-        []
-    );
-    const queryParameters = useRef<string>("");
-    const enabledFilters = useRef<
-        ("favorite" | "source" | "texts" | "sorts")[]
-    >([]);
-    const updateEnabledFilters = useCallback(
-        (
-            value: "favorite" | "source" | "texts" | "sorts",
-            enable: "enable" | "disable" = "enable"
-        ) => {
-            const currentList = enabledFilters.current;
-            if (enable === "enable") {
-                if (!currentList.includes(value)) {
-                    enabledFilters.current.push(value);
-                } else {
-                    enabledFilters.current = enabledFilters.current
-                        .filter((enabledItem) => enabledItem !== value)
-                        .concat([value]);
-                }
-            } else if (enable === "disable") {
-                if (currentList.includes(value)) {
-                    enabledFilters.current = enabledFilters.current.filter(
-                        (enabledItem) => enabledItem !== value
-                    );
-                }
-            }
-        },
-        []
-    );
-    const cacheContainer = useFileCaches();
     const queryFn = useCallback(
         ({ pageParam }: { pageParam: number }) =>
             isLocal
@@ -184,31 +185,12 @@ export default function MainPage({
             return currentPage - 1;
         },
     });
-    const updateFeedsCache = useCallback(
-        (
-            feedsList: ParsedFeedsDataType[],
-            cacheData: { cache: FeedsCache; lastPage: number }
-        ) => {
-            const { cache, lastPage } = cacheData;
-            const pageNumber = currentPage > lastPage ? currentPage : lastPage;
-            const currentPageList = cache[pageNumber];
-            if (currentPageList?.length > 0) {
-                const isListsIdentical = currentPageList.every(
-                    (feedData, index) => feedData.id === feedsList[index]?.id
-                );
-                if (!isListsIdentical) {
-                    cache[pageNumber] = currentPageList
-                        .slice(currentPageList.length)
-                        .concat(feedsList);
-                }
-            } else {
-                cache[pageNumber] = currentPageList
-                    .slice(currentPageList.length)
-                    .concat(feedsList);
-            }
-        },
-        [currentPage]
-    );
+    const { updateObserverElement } = useObserveElement({
+        callbackCondition: hasNextPage,
+        callback: () => setCurrentPage((oldPage) => oldPage + 1),
+    });
+
+    // display 관련
     const updateFeedsToDisplay = useCallback(
         (cache: FeedsCache) => {
             if (isMobileLayout) {
@@ -226,21 +208,94 @@ export default function MainPage({
         },
         [isMobileLayout, currentPage]
     );
-    const initializeFilteredCache = useCallback(() => {
-        cacheContainer.filtered.cache = {
-            ...cacheContainer.filtered.cache,
-            ...Array.from(
-                { length: Math.ceil(totalCount / 10) },
-                (_, k) => k + 1
-            ).reduce(
-                (result, pageIndex) => ({
-                    ...result,
-                    [pageIndex]: [],
-                }),
-                {}
-            ),
-        };
-    }, [cacheContainer, totalCount]);
+
+    // display 관련
+    const handleFeedsAndCache = useCallback(
+        (feedsList: ParsedFeedsDataType[]) => {
+            let cache: FeedsCache = cacheContainer.default.cache;
+            let lastPage: number = 1;
+            if (enabledFilters.current.length > 0) {
+                cache = cacheContainer.filtered.cache;
+                lastPage = cacheContainer.filtered.lastPage;
+            } else {
+                cache = cacheContainer.default.cache;
+                lastPage = cacheContainer.default.lastPage;
+            }
+            updateFeedsCache(feedsList, { cache, lastPage });
+            updateFeedsToDisplay(cache);
+        },
+        [updateFeedsCache, updateFeedsToDisplay, cacheContainer]
+    );
+
+    // 디스플레이 관련
+    const checkAndUpdateNewFeeds = useCallback(async () => {
+        try {
+            setRenewState(STATE_MESSAGE_STRINGS.proceed);
+            abortControllerRef.current = new AbortController();
+            const signal = abortControllerRef.current.signal;
+            const newFeedsRequestResult = await getDataFrom<
+                PageParamData | ErrorResponse
+            >(`/feeds/new?userId=${userId}`, { signal });
+            if (newFeedsRequestResult != null) {
+                switch (true) {
+                    case "data" in newFeedsRequestResult:
+                        const { data, count, updated } = newFeedsRequestResult;
+                        if (count !== totalCount) {
+                            setTotalCount(count);
+                        }
+                        if (updated !== 0) {
+                            setRenewState(
+                                updated + STATE_MESSAGE_STRINGS.added
+                            );
+                        } else {
+                            setRenewState(STATE_MESSAGE_STRINGS.end);
+                        }
+                        handleFeedsAndCache(data);
+                        break;
+                    case "error" in newFeedsRequestResult:
+                        setRenewState(
+                            STATE_MESSAGE_STRINGS[newFeedsRequestResult.error]
+                        );
+                        break;
+                    default:
+                        setRenewState(STATE_MESSAGE_STRINGS.no_change);
+                        break;
+                }
+            } else {
+                throw new Error("피드 갱신에 실패했습니다.");
+            }
+        } catch (error) {
+            console.error(error);
+        }
+    }, [getDataFrom, handleFeedsAndCache, totalCount, userId]);
+
+    // 필터링 관련
+    const updateEnabledFilters = useCallback(
+        (
+            value: "favorite" | "source" | "texts" | "sorts",
+            enable: "enable" | "disable" = "enable"
+        ) => {
+            const currentList = enabledFilters.current;
+            if (enable === "enable") {
+                if (!currentList.includes(value)) {
+                    enabledFilters.current.push(value);
+                } else {
+                    enabledFilters.current = enabledFilters.current
+                        .filter((enabledItem) => enabledItem !== value)
+                        .concat([value]);
+                }
+            } else if (enable === "disable") {
+                if (currentList.includes(value)) {
+                    enabledFilters.current = enabledFilters.current.filter(
+                        (enabledItem) => enabledItem !== value
+                    );
+                }
+            }
+        },
+        []
+    );
+
+    // 필터링 관련
     const filterBySources = useCallback(
         (newDisplay: SourceDisplayState) => {
             const lastDisplayState = JSON.stringify(sourceDisplayState);
@@ -288,6 +343,8 @@ export default function MainPage({
             initializeFilteredCache,
         ]
     );
+
+    // 필터링 관련
     const filterBySearchTexts = useCallback(
         (target: string, value: string) => {
             let lastPage: number = 1;
@@ -347,119 +404,8 @@ export default function MainPage({
             initializeFilteredCache,
         ]
     );
-    const handleFeedsAndCache = useCallback(
-        (feedsList: ParsedFeedsDataType[]) => {
-            let cache: FeedsCache = cacheContainer.default.cache;
-            let lastPage: number = 1;
-            if (enabledFilters.current.length > 0) {
-                cache = cacheContainer.filtered.cache;
-                lastPage = cacheContainer.filtered.lastPage;
-            } else {
-                cache = cacheContainer.default.cache;
-                lastPage = cacheContainer.default.lastPage;
-            }
-            updateFeedsCache(feedsList, { cache, lastPage });
-            updateFeedsToDisplay(cache);
-        },
-        [updateFeedsCache, updateFeedsToDisplay, cacheContainer]
-    );
-    const initializeCache = useCallback(
-        (indexList: number[], feedsList: ParsedFeedsDataType[]) => {
-            Object.entries(cacheContainer).forEach(([cacheKey, cacheData]) => {
-                indexList.forEach((pageIndex) => {
-                    if (cacheKey === "basic") {
-                        cacheData.cache[pageIndex] =
-                            pageIndex === 1 ? feedsList : [];
-                    } else {
-                        cacheData.cache[pageIndex] = [];
-                    }
-                });
-            });
-        },
-        [cacheContainer]
-    );
-    useEffect(() => {
-        if (feeds) {
-            const {
-                data,
-                count,
-            }: { data: ParsedFeedsDataType[]; count: number } =
-                JSON.parse(feeds);
-            setTotalCount(count);
-            const indexList = Array.from(
-                { length: Math.ceil(count / 10) },
-                (_, k) => k + 1
-            );
-            initializeCache(indexList, data);
-        }
-    }, [feeds, initializeCache]);
-    useEffect(() => {
-        queryParameters.current = generateSearchParameters({
-            ...(isFilterFavorite && {
-                favorites: isFilterFavorite,
-            }),
-            ...(Object.values(sourceDisplayState).includes(false) && {
-                displayOption: JSON.stringify(sourceDisplayState),
-            }),
-            ...(Object.values(searchTexts).some(
-                (searchText: string) => searchText.length >= 2
-            ) && { textOption: JSON.stringify(searchTexts) }),
-            ...(currentSort > 0 && { sortOption: currentSort }),
-        });
-    }, [isFilterFavorite, currentSort, searchTexts, sourceDisplayState]);
-    useEffect(() => {
-        if (storedFeed?.pages) {
-            const { data, count } = JSON.parse(
-                storedFeed.pages[storedFeed.pages.length - 1] ?? "{}"
-            );
-            if (count != null) setTotalCount(count);
-            if (data != null) {
-                handleFeedsAndCache(data);
-            }
-        }
-    }, [storedFeed, isMobileLayout, handleFeedsAndCache]);
 
-    const checkAndUpdateNewFeeds = useCallback(async () => {
-        try {
-            setRenewState(STATE_MESSAGE_STRINGS.proceed);
-            abortControllerRef.current = new AbortController();
-            const signal = abortControllerRef.current.signal;
-            const newFeedsRequestResult = await getDataFrom<
-                PageParamData | ErrorResponse
-            >(`/feeds/new?userId=${userId}`, { signal });
-            if (newFeedsRequestResult != null) {
-                switch (true) {
-                    case "data" in newFeedsRequestResult:
-                        const { data, count, updated } = newFeedsRequestResult;
-                        if (count !== totalCount) {
-                            setTotalCount(count);
-                        }
-                        if (updated !== 0) {
-                            setRenewState(
-                                updated + STATE_MESSAGE_STRINGS.added
-                            );
-                        } else {
-                            setRenewState(STATE_MESSAGE_STRINGS.end);
-                        }
-                        handleFeedsAndCache(data);
-                        break;
-                    case "error" in newFeedsRequestResult:
-                        setRenewState(
-                            STATE_MESSAGE_STRINGS[newFeedsRequestResult.error]
-                        );
-                        break;
-                    default:
-                        setRenewState(STATE_MESSAGE_STRINGS.no_change);
-                        break;
-                }
-            } else {
-                throw new Error("피드 갱신에 실패했습니다.");
-            }
-        } catch (error) {
-            console.error(error);
-        }
-    }, [getDataFrom, handleFeedsAndCache, totalCount, userId]);
-
+    // 필터링 관련
     const filterBySort = useCallback(
         (stateStringArray: string[]) => (stateString: string) => {
             let lastPage: number = 1;
@@ -501,6 +447,7 @@ export default function MainPage({
         ]
     );
 
+    // 필터링 관련
     const filterFavorites = useCallback(() => {
         setIsFilterFavorite(!isFilterFavorite);
         let lastPage: number = 1;
@@ -526,56 +473,54 @@ export default function MainPage({
         setCurrentPage(lastPage);
     }, [
         cacheContainer,
-        initializeFilteredCache,
-        updateEnabledFilters,
         isFilterFavorite,
         isMobileLayout,
+        updateEnabledFilters,
+        initializeFilteredCache,
     ]);
 
-    const updateObserverElement = (element: HTMLDivElement) => {
-        setObserverElement(element);
-    };
-
-    const updateCurrentPage = (value: number | ((value: number) => number)) => {
-        setCurrentPage(value);
-    };
-
-    const detectIfMobileLayout = useCallback(() => {
-        if (document.documentElement.offsetWidth >= 768) {
-            setIsMobileLayout(false);
-        } else {
-            setIsMobileLayout(true);
+    useEffect(() => {
+        if (feeds) {
+            const {
+                data,
+                count,
+            }: { data: ParsedFeedsDataType[]; count: number } =
+                JSON.parse(feeds);
+            setTotalCount(count);
+            const indexList = Array.from(
+                { length: Math.ceil(count / 10) },
+                (_, k) => k + 1
+            );
+            initializeCache(indexList, data);
         }
-    }, []);
-
-    useResizeEvent(detectIfMobileLayout, true, [detectIfMobileLayout]);
+    }, [feeds, initializeCache]);
 
     useEffect(() => {
-        if (observerElement != null) {
-            const observerOption: IntersectionObserverInit = {
-                threshold: 0.5,
-            };
-            const observerCallback: IntersectionObserverCallback = (
-                entries: IntersectionObserverEntry[]
-            ) => {
-                entries.forEach((entry: IntersectionObserverEntry) => {
-                    if (entry.isIntersecting) {
-                        if (hasNextPage) {
-                            setCurrentPage(
-                                (previousValue) => previousValue + 1
-                            );
-                        }
-                    }
-                });
-            };
-            const observer = new IntersectionObserver(
-                observerCallback,
-                observerOption
+        queryParameters.current = generateSearchParameters({
+            ...(isFilterFavorite && {
+                favorites: isFilterFavorite,
+            }),
+            ...(Object.values(sourceDisplayState).includes(false) && {
+                displayOption: JSON.stringify(sourceDisplayState),
+            }),
+            ...(Object.values(searchTexts).some(
+                (searchText: string) => searchText.length >= 2
+            ) && { textOption: JSON.stringify(searchTexts) }),
+            ...(currentSort > 0 && { sortOption: currentSort }),
+        });
+    }, [isFilterFavorite, currentSort, searchTexts, sourceDisplayState]);
+
+    useEffect(() => {
+        if (storedFeed?.pages) {
+            const { data, count } = JSON.parse(
+                storedFeed.pages[storedFeed.pages.length - 1] ?? "{}"
             );
-            observer.observe(observerElement);
-            return () => observer.unobserve(observerElement);
+            if (count != null) setTotalCount(count);
+            if (data != null) {
+                handleFeedsAndCache(data);
+            }
         }
-    }, [observerElement, hasNextPage, fetchNextPage]);
+    }, [storedFeed, isMobileLayout, handleFeedsAndCache]);
 
     useEffect(() => {
         if (abortControllerRef.current) {
